requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
Flask-Swagger-UI==5.21.0
PyYAML==6.0.1
gunicorn==21.2.0
